# Web/PatientDetails/Models/PatientDetailsODSheet.cs
namespace PatientDetails.Models
{
    public class PatientDetailsODSheet
    {
        public string? patientName { get; set; }
        public string? address { get; set; }
        public string? dob { get; set; }
        public string? email { get; set; }
        public string? phone { get; set; }
        public string? drugId { get; set; }
        public string? drugName { get; set; }
    }
}